+++ UpgradeManagers/Regeneration.cs
@@ -1,43 +1,18 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using BepInEx.Configuration;
-using HarmonyLib;
 using UnityEngine;
 using static HarmonyLib.AccessTools;
-using Object = UnityEngine.Object;
 
 namespace SLRUpgradePack.UpgradeManagers;
 
-public class RegenerationComponent : MonoBehaviour {
-    internal PlayerAvatar player;
-    private float pendingHealing = 0;
-    private readonly FieldRef<PlayerHealth, int> _healthRef = FieldRefAccess<PlayerHealth, int>("health");
-
-    private void Start() {
-        SLRUpgradePack.Logger.LogInfo($"{SemiFunc.PlayerGetName(player)} is regenerating");
-    }
-
-    private void Update() {
-        var regenerationUpgrade = SLRUpgradePack.RegenerationUpgradeInstance;
-
-        if (!ValuableDirector.instance.setupComplete) return;
-
-        if (!regenerationUpgrade.UpgradeEnabled.Value || regenerationUpgrade.UpgradeRegister.GetLevel(player) == 0 || _healthRef.Invoke(player.playerHealth) == 0) return;
-
-        pendingHealing +=
-            regenerationUpgrade.Calculate(regenerationUpgrade.BaseHealing.Value * Time.deltaTime, player,
-                regenerationUpgrade.UpgradeRegister.GetLevel(player));
-
-        if (pendingHealing >= 1) {
-            player.playerHealth.Heal((int)Math.Floor(pendingHealing), false);
-            pendingHealing -= Mathf.Floor(pendingHealing);
-        }
-    }
-}
-
 public class RegenerationUpgrade : UpgradeBase<float> {
     public ConfigEntry<float> BaseHealing { get; protected set; }
-    internal Dictionary<string, RegenerationComponent> Regenerations { get; set; } = new();
+    internal static Dictionary<string, PlayerAvatar> RegeneratingPlayers { get; } = new();
+    internal static Dictionary<string, float> pendingHealings { get; } = new();
+    private readonly FieldRef<PlayerHealth, int> _healthRef = FieldRefAccess<PlayerHealth, int>("health");
+    private readonly FieldRef<ValuableDirector, bool> _setupCompleteRef = FieldRefAccess<ValuableDirector, bool>("setupComplete");
 
     public RegenerationUpgrade(bool enabled, float upgradeAmount, bool exponential, float exponentialAmount,
         ConfigFile config, AssetBundle assetBundle, float baseHealing, float priceMultiplier) : base("Regeneration", "assets/repo/mods/resources/items/items/item upgrade regeneration lib.asset", enabled, upgradeAmount,
@@ -46,21 +21,41 @@ public class RegenerationUpgrade : UpgradeBase<float> {
     }
 
     public override float Calculate(float value, PlayerAvatar player, int level) => DefaultCalculateFloatIncrease(this, "Regeneration", value, player, level);
-}
 
-[HarmonyPatch(typeof(LevelGenerator))]
-public class LevelGeneratorRegenerationPatch {
-    [HarmonyPatch("GenerateDone")]
-    [HarmonyPostfix]
-    private static void GenerateDonePostfix() {
-        var regenerationUpgradeInstance = SLRUpgradePack.RegenerationUpgradeInstance;
-        var player = PlayerController.instance.playerAvatarScript;
-        if (player == null) return;
-        SLRUpgradePack.Logger.LogInfo($"Adding regeneration component to {SemiFunc.PlayerGetName(player)} ({SemiFunc.PlayerGetSteamID(player)})");
-        if (regenerationUpgradeInstance.Regenerations.TryGetValue(SemiFunc.PlayerGetSteamID(player), out var regenerationComponent) && regenerationComponent != null) Object.Destroy(regenerationComponent);
+    internal override void InitUpgrade(PlayerAvatar player, int level) {
+        base.InitUpgrade(player, level);
+        if (!SemiFunc.IsMasterClientOrSingleplayer()) return; // host handles all regeneration
 
-        regenerationComponent = player.gameObject.AddComponent<RegenerationComponent>();
-        regenerationComponent.player = player;
-        regenerationUpgradeInstance.Regenerations[SemiFunc.PlayerGetSteamID(player)] = regenerationComponent;
+        foreach (var missingPlayer in RegeneratingPlayers.Keys.Where(x => !UpgradeRegister.PlayerDictionary.ContainsKey(x)).ToList()) {
+            RegeneratingPlayers.Remove(missingPlayer);
+            pendingHealings.Remove(missingPlayer);
+        }
+
+        RegeneratingPlayers[SemiFunc.PlayerGetSteamID(player)] = player;
+
+        if (!SLRUpgradePack.Instance.Actions.ContainsKey("Regeneration"))
+            SLRUpgradePack.Instance.Actions.Add("Regeneration", RegenerationUpdateAction);
+    }
+
+    private void RegenerationUpdateAction() {
+        foreach (var player in RegeneratingPlayers.Values) {
+            var playerID = SemiFunc.PlayerGetSteamID(player);
+            if (!_setupCompleteRef.Invoke(ValuableDirector.instance)) continue;
+
+            if (!UpgradeEnabled.Value || UpgradeRegister.GetLevel(player) == 0 || _healthRef.Invoke(player.playerHealth) == 0) continue;
+
+            pendingHealings.TryGetValue(playerID, out var pendingHealing);
+
+            pendingHealing +=
+                Calculate(BaseHealing.Value * Time.deltaTime, player,
+                    UpgradeRegister.GetLevel(player));
+
+            if (pendingHealing >= 1) {
+                player.playerHealth.HealOther((int)Math.Floor(pendingHealing), false);
+                pendingHealing -= Mathf.Floor(pendingHealing);
+            }
+
+            pendingHealings[playerID] = pendingHealing;
+        }
     }
 }